Fix recording analytics filtering, pagination and table data download
The issue was caused by filter params not being passed through to the
pagination filters and csv download. Fixed by ensuring these values are
provided to set_url which will allow it to be passed along as needed,
and added additional handling to ensure the filter form stays filtered
on subsequent pagination changes.