fix: processing of recording metadata from analytics
Though the data was available, it was not properly considered when updating the status of the recording. Also since the recording state was never set in the CREATE log record, and only in the SUMMARY, this was never properly reflected in the table.

Instead of doing a join just for reporting purposes, now it will update the recorded flag via the task, and should sync the recordings report with the proper status for sessions that are recorded, but not yet available as 'Processing'.